requirements.txt
opencv-python
numpy
ultralytics
tensorflow
Pillow